+++ src/assembly/mod.rs
@@ -2,3 +2,4 @@ pub mod explication;
 mod explication_context;
 mod explication_explicator;
 pub mod parser;
+mod explication_util;
